@@ -1,13 +1,4 @@
-const javascript = require('./rules/javascript');
-const lit = require('./rules/lit');
-const react = require('./rules/react');
-const typescript = require('./rules/typescript');
-const preset = require('./preset');
-
-module.exports = {
-  javascript,
-  lit,
-  preset,
-  react,
-  typescript,
-};
+export * as javascript from './rules/javascript.js';
+export * as typescript from './rules/typescript.js';
+// import lit from './rules/lit.js';
+// import react from './rules/react.js';
